Add force_destroy and importer to index resource.
Closes https://github.com/phillbaker/terraform-provider-elasticsearch/issues/46.

+++ resource_elasticsearch_index.go
@@ -2,8 +2,12 @@ package main
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
+	"log"
 
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 
 	elastic7 "github.com/olivere/elastic/v7"
 	elastic5 "gopkg.in/olivere/elastic.v5"
@@ -11,13 +15,13 @@ import (
 )
 
 var (
-	staticConfigKeys = []string{
+	staticSettingsKeys = []string{
 		"number_of_shards",
 		"codec",
 		"routing_partition_size",
 		"load_fixed_bitset_filters_eagerly",
 	}
-	dynamicConfigKeys = []string{
+	dynamicsSettingsKeys = []string{
 		"number_of_replicas",
 		"auto_expand_replicas",
 		"refresh_interval",
@@ -26,78 +30,81 @@ var (
 		//"max_rescore_window"
 		//...
 	}
-	configKeys = append(staticConfigKeys, dynamicConfigKeys...)
+	settingsKeys = append(staticSettingsKeys, dynamicsSettingsKeys...)
 )
 
 var (
 	configSchema = map[string]*schema.Schema{
-		// Static settings that can only be set on creation
-		"name": &schema.Schema{
+		"name": {
 			Type:        schema.TypeString,
 			Description: "Name of the index to create",
 			ForceNew:    true,
 			Required:    true,
 		},
-		"number_of_shards": &schema.Schema{
-			Type:        schema.TypeInt,
-			Description: "Number of shards for the index",
-			ForceNew:    true, // shards can only set upon creation
-			Default:     1,
+		"force_destroy": {
+			Type:        schema.TypeBool,
+			Description: "A boolean that indicates that the index should be deleted even if it contains documents.",
+			Default:     false,
 			Optional:    true,
 		},
-		// "check_on_startup": &schema.Schema{
-		// 	Type:     schema.TypeString, // false,checksum,true
-		// 	ForceNew: true,
-		// 	// Default:  "false",
-		// 	Optional: true,
-		// },
-		"routing_partition_size": &schema.Schema{
+		// Static settings that can only be set on creation
+		"number_of_shards": {
+			Type:        schema.TypeString,
+			Description: "Number of shards for the index",
+			ForceNew:    true, // shards can only set upon creation
+			Default:     "1",
+			Optional:    true,
+		},
+		"routing_partition_size": {
 			Type:     schema.TypeInt,
 			ForceNew: true, // shards can only set upon creation
-			// Default:  1,
 			Optional: true,
 		},
-		"load_fixed_bitset_filters_eagerly": &schema.Schema{
+		"load_fixed_bitset_filters_eagerly": {
 			Type:     schema.TypeBool,
-			ForceNew: true, // false,checksum,true
-			// Default:  true,
+			ForceNew: true,
 			Optional: true,
 		},
-		"codec": &schema.Schema{
+		"codec": {
 			Type:     schema.TypeString,
 			ForceNew: true,
-			// Default:  "default",
 			Optional: true,
 		},
 		// Dynamic settings that can be changed at runtime
-		"number_of_replicas": &schema.Schema{
-			Type:        schema.TypeInt,
+		"number_of_replicas": {
+			Type:        schema.TypeString,
 			Description: "Number of shard replicas",
-			// Default:  1,
-			Optional: true,
+			Optional:    true,
 		},
-		"auto_expand_replicas": &schema.Schema{
+		"auto_expand_replicas": {
 			Type:        schema.TypeString, // 0-5 OR 0-all
 			Description: "Set the number of replicas to the node count in the cluster",
-			// Default:  "false",
+			Optional:    true,
+		},
+		"refresh_interval": {
+			Type:     schema.TypeString, // -1 to disable
 			Optional: true,
 		},
-		"refresh_interval": &schema.Schema{
-			Type: schema.TypeString, // -1 to disable
-			// Default:  "1s",
+		// Other attributes
+		"mappings": {
+			Type:     schema.TypeString,
 			Optional: true,
+			// In order to not handle complexities of field mapping updates, updates
+			// are not allowed via this provider. See
+			// https://www.elastic.co/guide/en/elasticsearch/reference/6.8/indices-put-mapping.html#updating-field-mappings.
+			ForceNew:     true,
+			ValidateFunc: validation.ValidateJsonString,
+		},
+		"aliases": {
+			Type:     schema.TypeString,
+			Optional: true,
+			// In order to not handle the separate endpoint of alias updates, updates
+			// are not allowed via this provider currently.
+			ForceNew:     true,
+			ValidateFunc: validation.ValidateJsonString,
 		},
 	}
 )
-
-// this is a check to ensure consistency when new settings are supported.
-// It ensures the listed config keys match the schema as new keys must
-// be registered in both places
-func init() {
-	if len(configKeys) != len(configSchema)-1 {
-		panic("declared keys do not match the schema")
-	}
-}
 
 func resourceElasticsearchIndex() *schema.Resource {
 	return &schema.Resource{
@@ -106,37 +113,55 @@ func resourceElasticsearchIndex() *schema.Resource {
 		Update: resourceElasticsearchIndexUpdate,
 		Delete: resourceElasticsearchIndexDelete,
 		Schema: configSchema,
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
 	}
 }
 
 func resourceElasticsearchIndexCreate(d *schema.ResourceData, meta interface{}) error {
 	var (
 		name     = d.Get("name").(string)
-		settings = settingsFromResourceData(d)
+		settings = settingsFromIndexResourceData(d)
 		body     = make(map[string]interface{})
 		ctx      = context.Background()
 		err      error
 	)
-
 	if len(settings) > 0 {
 		body["settings"] = settings
+
+		if aliasJson, ok := d.GetOk("aliases"); ok {
+			var aliases map[string]interface{}
+			bytes := []byte(aliasJson.(string))
+			err = json.Unmarshal(bytes, &aliases)
+			if err != nil {
+				return fmt.Errorf("fail to unmarshal: %v", err)
+			}
+			body["aliases"] = aliases
+		}
+
+		if mappingsJson, ok := d.GetOk("mappings"); ok {
+			var mappings map[string]interface{}
+			bytes := []byte(mappingsJson.(string))
+			err = json.Unmarshal(bytes, &mappings)
+			if err != nil {
+				return fmt.Errorf("fail to unmarshal: %v", err)
+			}
+			body["mappings"] = mappings
+		}
 	}
 
-	switch meta.(type) {
+	switch client := meta.(type) {
 	case *elastic7.Client:
-		client := meta.(*elastic7.Client)
 		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
 
 	case *elastic6.Client:
-		client := meta.(*elastic6.Client)
 		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
 
 	default:
-		client := meta.(*elastic5.Client)
-		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
+		elastic5Client := meta.(*elastic5.Client)
+		_, err = elastic5Client.CreateIndex(name).BodyJson(body).Do(ctx)
 	}
-
-	// return err
 
 	if err == nil {
 		// Let terraform know the resource was created
@@ -146,9 +171,9 @@ func resourceElasticsearchIndexCreate(d *schema.ResourceData, meta interface{}) 
 	return err
 }
 
-func settingsFromResourceData(d *schema.ResourceData) map[string]interface{} {
+func settingsFromIndexResourceData(d *schema.ResourceData) map[string]interface{} {
 	settings := make(map[string]interface{})
-	for _, key := range configKeys {
+	for _, key := range settingsKeys {
 		if raw, ok := d.GetOk(key); ok {
 			settings[key] = raw
 		}
@@ -156,72 +181,77 @@ func settingsFromResourceData(d *schema.ResourceData) map[string]interface{} {
 	return settings
 }
 
-func resourceDataFromSettings(settings map[string]interface{}, d *schema.ResourceData) {
-	for _, key := range configKeys {
-		if val, ok := settings[key]; ok {
-			d.Set(key, val)
+func indexResourceDataFromSettings(settings map[string]interface{}, d *schema.ResourceData) {
+	for _, key := range settingsKeys {
+		err := d.Set(key, settings[key])
+		if err != nil {
+			log.Printf("[INFO] indexResourceDataFromSettings: %+v", err)
 		}
 	}
-	// if raw, ok := d.GetOk("check_on_startup"); ok {
-	// 	settings["shard.check_on_startup"] = raw.(string)
-	// }
-}
-
-func elasticIndexCreate(meta interface{}, name string, settings, mappings map[string]interface{}) error {
-	var (
-		body = make(map[string]interface{})
-		ctx  = context.Background()
-		err  error
-	)
-
-	if len(settings) > 0 {
-		body["settings"] = settings
-	}
-
-	switch meta.(type) {
-	case *elastic7.Client:
-		client := meta.(*elastic7.Client)
-		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
-
-	case *elastic6.Client:
-		client := meta.(*elastic6.Client)
-		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
-
-	default:
-		client := meta.(*elastic5.Client)
-		_, err = client.CreateIndex(name).BodyJson(body).Do(ctx)
-	}
-
-	return err
 }
 
 func resourceElasticsearchIndexDelete(d *schema.ResourceData, meta interface{}) error {
 	var (
-		name = d.Get("name").(string)
+		name = d.Id()
 		ctx  = context.Background()
 		err  error
 	)
 
-	switch meta.(type) {
+	// check to see if there are documents in the index
+	allowed := allowIndexDestroy(name, d, meta)
+	if !allowed {
+		return fmt.Errorf("There are documents in the index (or the index could not be , set force_destroy to true to allow destroying.")
+	}
+
+	switch client := meta.(type) {
 	case *elastic7.Client:
-		client := meta.(*elastic7.Client)
 		_, err = client.DeleteIndex(name).Do(ctx)
 
 	case *elastic6.Client:
-		client := meta.(*elastic6.Client)
 		_, err = client.DeleteIndex(name).Do(ctx)
 
 	default:
-		client := meta.(*elastic5.Client)
-		_, err = client.DeleteIndex(name).Do(ctx)
+		elastic5Client := meta.(*elastic5.Client)
+		_, err = elastic5Client.DeleteIndex(name).Do(ctx)
 	}
 
 	return err
 }
 
+func allowIndexDestroy(indexName string, d *schema.ResourceData, meta interface{}) bool {
+	force := d.Get("force_destroy").(bool)
+
+	var (
+		ctx   = context.Background()
+		count int64
+		err   error
+	)
+	switch client := meta.(type) {
+	case *elastic7.Client:
+		count, err = client.Count(indexName).Do(ctx)
+
+	case *elastic6.Client:
+		count, err = client.Count(indexName).Do(ctx)
+
+	default:
+		elastic5Client := meta.(*elastic5.Client)
+		count, err = elastic5Client.Count(indexName).Do(ctx)
+	}
+
+	if err != nil {
+		log.Printf("[INFO] allowIndexDestroy: %+v", err)
+		return false
+	}
+
+	if count > 0 && !force {
+		return false
+	}
+	return true
+}
+
 func resourceElasticsearchIndexUpdate(d *schema.ResourceData, meta interface{}) error {
 	settings := make(map[string]interface{})
-	for _, key := range configKeys {
+	for _, key := range settingsKeys {
 		if d.HasChange(key) {
 			settings[key] = d.Get(key)
 		}
@@ -231,23 +261,21 @@ func resourceElasticsearchIndexUpdate(d *schema.ResourceData, meta interface{}) 
 	}
 
 	var (
-		name = d.Get("name").(string)
+		name = d.Id()
 		ctx  = context.Background()
 		err  error
 	)
 
-	switch meta.(type) {
+	switch client := meta.(type) {
 	case *elastic7.Client:
-		client := meta.(*elastic7.Client)
 		_, err = client.IndexPutSettings(name).BodyJson(body).Do(ctx)
 
 	case *elastic6.Client:
-		client := meta.(*elastic6.Client)
 		_, err = client.IndexPutSettings(name).BodyJson(body).Do(ctx)
 
 	default:
-		client := meta.(*elastic5.Client)
-		_, err = client.IndexPutSettings(name).BodyJson(body).Do(ctx)
+		elastic5Client := meta.(*elastic5.Client)
+		_, err = elastic5Client.IndexPutSettings(name).BodyJson(body).Do(ctx)
 	}
 
 	if err == nil {
@@ -258,50 +286,44 @@ func resourceElasticsearchIndexUpdate(d *schema.ResourceData, meta interface{}) 
 
 func resourceElasticsearchIndexRead(d *schema.ResourceData, meta interface{}) error {
 	var (
-		name     = d.Get("name").(string)
+		name     = d.Id()
 		ctx      = context.Background()
 		settings map[string]interface{}
 	)
 
 	// The logic is repeated strictly becuase of the types
-	switch meta.(type) {
+	switch client := meta.(type) {
 	case *elastic7.Client:
-		client := meta.(*elastic7.Client)
 		r, err := client.IndexGet(name).Do(ctx)
 		if err != nil {
 			return err
 		}
 
 		resp := r[name]
-		// aliases = resp.Aliases
 		settings = resp.Settings["index"].(map[string]interface{})
 
 	case *elastic6.Client:
-		client := meta.(*elastic6.Client)
 		r, err := client.IndexGet(name).Do(ctx)
 		if err != nil {
 			return err
 		}
 
 		resp := r[name]
-		// aliases = resp.Aliases
-		settings = resp.Settings
+		settings = resp.Settings["index"].(map[string]interface{})
 
 	default:
-		client := meta.(*elastic5.Client)
-		r, err := client.IndexGet(name).Do(ctx)
+		elastic5Client := meta.(*elastic5.Client)
+		r, err := elastic5Client.IndexGet(name).Do(ctx)
 		if err != nil {
 			return err
 		}
 
 		resp := r[name]
-		// aliases = resp.Aliases
-		settings = resp.Settings
+		settings = resp.Settings["index"].(map[string]interface{})
 
 	}
-
 	d.Set("name", name)
-	resourceDataFromSettings(settings, d)
+	indexResourceDataFromSettings(settings, d)
 
 	return nil
 }
